requirements.txt
protobuf >= 3.14
grpcio >= 1.34
grpcio-tools >= 1.34
google-auth >= 1.23
requests >= 2.25
wheel >= 0.36.2
pyaudio >= 0.2.11
flask >= 1.1.2
flask_classful >= 0.14.2